+++ v2/models/application_rule_settings.go
@@ -3,5 +3,5 @@ package models
 const ApplicationRuleSettingsPath = "application/rule-settings/"
 
 type ApplicationRuleSettings struct {
-	Fosid *float64 `json:"fosid,omitempty"`
+	Fosid *int64 `json:"fosid,omitempty"`
 }
