+++ v2/models/dnsfilter_domain_filter.go
@@ -5,14 +5,14 @@ const DnsfilterDomainFilterPath = "dnsfilter/domain-filter/"
 type DnsfilterDomainFilter struct {
 	Comment *string                         `json:"comment,omitempty"`
 	Entries *[]DnsfilterDomainFilterEntries `json:"entries,omitempty"`
-	Fosid   *float64                        `json:"fosid,omitempty"`
+	Fosid   *int64                          `json:"fosid,omitempty"`
 	Name    *string                         `json:"name,omitempty"`
 }
 
 type DnsfilterDomainFilterEntries struct {
-	Action *string  `json:"action,omitempty"`
-	Domain *string  `json:"domain,omitempty"`
-	Id     *float64 `json:"id,omitempty"`
-	Status *string  `json:"status,omitempty"`
-	Type   *string  `json:"type,omitempty"`
+	Action *string `json:"action,omitempty"`
+	Domain *string `json:"domain,omitempty"`
+	Id     *int64  `json:"id,omitempty"`
+	Status *string `json:"status,omitempty"`
+	Type   *string `json:"type,omitempty"`
 }
