+++ v2/models/extender_controller_extender.go
@@ -9,17 +9,17 @@ type ExtenderControllerExtender struct {
 	Allowaccess                 *string                                       `json:"allowaccess,omitempty"`
 	AtDialScript                *string                                       `json:"at-dial-script,omitempty"`
 	Authorized                  *string                                       `json:"authorized,omitempty"`
-	BandwidthLimit              *float64                                      `json:"bandwidth-limit,omitempty"`
-	BillingStartDay             *float64                                      `json:"billing-start-day,omitempty"`
+	BandwidthLimit              *int64                                        `json:"bandwidth-limit,omitempty"`
+	BillingStartDay             *int64                                        `json:"billing-start-day,omitempty"`
 	CdmaAaaSpi                  *string                                       `json:"cdma-aaa-spi,omitempty"`
 	CdmaHaSpi                   *string                                       `json:"cdma-ha-spi,omitempty"`
 	CdmaNai                     *string                                       `json:"cdma-nai,omitempty"`
-	ConnStatus                  *float64                                      `json:"conn-status,omitempty"`
+	ConnStatus                  *int64                                        `json:"conn-status,omitempty"`
 	ControllerReport            *[]ExtenderControllerExtenderControllerReport `json:"controller-report,omitempty"`
 	Description                 *string                                       `json:"description,omitempty"`
-	DeviceId                    *float64                                      `json:"device-id,omitempty"`
+	DeviceId                    *int64                                        `json:"device-id,omitempty"`
 	DialMode                    *string                                       `json:"dial-mode,omitempty"`
-	DialStatus                  *float64                                      `json:"dial-status,omitempty"`
+	DialStatus                  *int64                                        `json:"dial-status,omitempty"`
 	EnforceBandwidth            *string                                       `json:"enforce-bandwidth,omitempty"`
 	ExtName                     *string                                       `json:"ext-name,omitempty"`
 	ExtensionType               *string                                       `json:"extension-type,omitempty"`
@@ -45,14 +45,14 @@ type ExtenderControllerExtender struct {
 	PppUsername                 *string                                       `json:"ppp-username,omitempty"`
 	PrimaryHa                   *string                                       `json:"primary-ha,omitempty"`
 	Profile                     *string                                       `json:"profile,omitempty"`
-	QuotaLimitMb                *float64                                      `json:"quota-limit-mb,omitempty"`
+	QuotaLimitMb                *int64                                        `json:"quota-limit-mb,omitempty"`
 	Redial                      *string                                       `json:"redial,omitempty"`
 	RedundantIntf               *string                                       `json:"redundant-intf,omitempty"`
 	Roaming                     *string                                       `json:"roaming,omitempty"`
 	Role                        *string                                       `json:"role,omitempty"`
 	SecondaryHa                 *string                                       `json:"secondary-ha,omitempty"`
 	SimPin                      *string                                       `json:"sim-pin,omitempty"`
-	Vdom                        *float64                                      `json:"vdom,omitempty"`
+	Vdom                        *int64                                        `json:"vdom,omitempty"`
 	WanExtension                *[]ExtenderControllerExtenderWanExtension     `json:"wan-extension,omitempty"`
 	WimaxAuthProtocol           *string                                       `json:"wimax-auth-protocol,omitempty"`
 	WimaxCarrier                *string                                       `json:"wimax-carrier,omitempty"`
@@ -60,14 +60,14 @@ type ExtenderControllerExtender struct {
 }
 
 type ExtenderControllerExtenderControllerReport struct {
-	Interval        *float64 `json:"interval,omitempty"`
-	SignalThreshold *float64 `json:"signal-threshold,omitempty"`
-	Status          *string  `json:"status,omitempty"`
+	Interval        *int64  `json:"interval,omitempty"`
+	SignalThreshold *int64  `json:"signal-threshold,omitempty"`
+	Status          *string `json:"status,omitempty"`
 }
 
 type ExtenderControllerExtenderModem1 struct {
 	AutoSwitch       *[]ExtenderControllerExtenderModem1AutoSwitch `json:"auto-switch,omitempty"`
-	ConnStatus       *float64                                      `json:"conn-status,omitempty"`
+	ConnStatus       *int64                                        `json:"conn-status,omitempty"`
 	DefaultSim       *string                                       `json:"default-sim,omitempty"`
 	Gps              *string                                       `json:"gps,omitempty"`
 	Ifname           *string                                       `json:"ifname,omitempty"`
@@ -81,19 +81,19 @@ type ExtenderControllerExtenderModem1 struct {
 }
 
 type ExtenderControllerExtenderModem1AutoSwitch struct {
-	Dataplan            *string  `json:"dataplan,omitempty"`
-	Disconnect          *string  `json:"disconnect,omitempty"`
-	DisconnectPeriod    *float64 `json:"disconnect-period,omitempty"`
-	DisconnectThreshold *float64 `json:"disconnect-threshold,omitempty"`
-	Signal              *string  `json:"signal,omitempty"`
-	SwitchBack          *string  `json:"switch-back,omitempty"`
-	SwitchBackTime      *string  `json:"switch-back-time,omitempty"`
-	SwitchBackTimer     *float64 `json:"switch-back-timer,omitempty"`
+	Dataplan            *string `json:"dataplan,omitempty"`
+	Disconnect          *string `json:"disconnect,omitempty"`
+	DisconnectPeriod    *int64  `json:"disconnect-period,omitempty"`
+	DisconnectThreshold *int64  `json:"disconnect-threshold,omitempty"`
+	Signal              *string `json:"signal,omitempty"`
+	SwitchBack          *string `json:"switch-back,omitempty"`
+	SwitchBackTime      *string `json:"switch-back-time,omitempty"`
+	SwitchBackTimer     *int64  `json:"switch-back-timer,omitempty"`
 }
 
 type ExtenderControllerExtenderModem2 struct {
 	AutoSwitch       *[]ExtenderControllerExtenderModem2AutoSwitch `json:"auto-switch,omitempty"`
-	ConnStatus       *float64                                      `json:"conn-status,omitempty"`
+	ConnStatus       *int64                                        `json:"conn-status,omitempty"`
 	DefaultSim       *string                                       `json:"default-sim,omitempty"`
 	Gps              *string                                       `json:"gps,omitempty"`
 	Ifname           *string                                       `json:"ifname,omitempty"`
@@ -107,14 +107,14 @@ type ExtenderControllerExtenderModem2 struct {
 }
 
 type ExtenderControllerExtenderModem2AutoSwitch struct {
-	Dataplan            *string  `json:"dataplan,omitempty"`
-	Disconnect          *string  `json:"disconnect,omitempty"`
-	DisconnectPeriod    *float64 `json:"disconnect-period,omitempty"`
-	DisconnectThreshold *float64 `json:"disconnect-threshold,omitempty"`
-	Signal              *string  `json:"signal,omitempty"`
-	SwitchBack          *string  `json:"switch-back,omitempty"`
-	SwitchBackTime      *string  `json:"switch-back-time,omitempty"`
-	SwitchBackTimer     *float64 `json:"switch-back-timer,omitempty"`
+	Dataplan            *string `json:"dataplan,omitempty"`
+	Disconnect          *string `json:"disconnect,omitempty"`
+	DisconnectPeriod    *int64  `json:"disconnect-period,omitempty"`
+	DisconnectThreshold *int64  `json:"disconnect-threshold,omitempty"`
+	Signal              *string `json:"signal,omitempty"`
+	SwitchBack          *string `json:"switch-back,omitempty"`
+	SwitchBackTime      *string `json:"switch-back-time,omitempty"`
+	SwitchBackTimer     *int64  `json:"switch-back-timer,omitempty"`
 }
 
 type ExtenderControllerExtenderWanExtension struct {
